renae CalcIndexDataSize -> CalcVectorDataSize
move components tests from test_common to test_components
